+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 **Version v2 (master)**
 
+- Parallelize scaling
+- Scale STATPOP to both projected household and person-level control totals
 - Add canton ID to households
 - Add car_passenger mode
-- Scale STATPOP to projected household-level control totals
 - Fix municipality type and canton imputation
 - Write out detailed information from MZ for PT trips
 - Fix education location